cloud_functions/data_transformation/requirements.txt
functions-framework==3.0.0
google-cloud-bigquery==2.34.4
